refactor(131+132): split GrooveURLCodec and PerformanceTracker into focused modules
Issue #131 - GrooveURLCodec.ts (599 → 388 lines, -35%):
- New src/core/urlCodecSchemas.ts: Zod schemas, VALIDATION, shared constants
- New src/core/stickingCodec.ts: sticking encode/decode functions
- GrooveURLCodec.ts now imports from those and keeps encode/decode/URL logic

Issue #132 - PerformanceTracker.ts (559 → 334 lines, -40%):
- New src/midi/BPMEstimator.ts: EWMA BPM estimation class (Spec Section 7)
- New src/midi/performanceGrader.ts: pure timing accuracy calculation functions
- PerformanceTracker.ts delegates to BPMEstimator and imported pure functions

All 239 tests pass, TypeScript clean.
