+++ src/core/GrooveURLCodec.ts
@@ -8,7 +8,6 @@
  * ?TimeSig=4/4&Div=16&Tempo=120&Measures=1&H=|x-x-x-x-|&S=|----o---|&K=|o-------|
  */
 
-import { z } from 'zod';
 import {
   GrooveData,
   DrumVoice,
@@ -19,110 +18,23 @@ import {
   getFlattenedNotes,
 } from '../types';
 import { URL_TAB_CHARS, URL_VOICE_GROUPS } from './DrumVoiceConfig';
-
-/** Validation constraints for URL parameters */
-const VALIDATION = {
-  TEMPO: { MIN: 20, MAX: 400, DEFAULT: 120 },
-  SWING: { MIN: 0, MAX: 100, DEFAULT: 0 },
-  MEASURES: { MIN: 1, MAX: 32, DEFAULT: 1 },
-  BEATS: { MIN: 1, MAX: 16, DEFAULT: 4 },
-  TITLE_MAX_LENGTH: 200,
-  AUTHOR_MAX_LENGTH: 100,
-  COMMENTS_MAX_LENGTH: 1000,
-  PATTERN_MAX_LENGTH: 2000,
-  STICKING_MAX_LENGTH: 600, // 16 measures * 48 subdivisions max = 768, allow some headroom
-} as const;
-
-/**
- * Sticking encoding: maps StickingValue to a single URL-safe character.
- * null → '-', 'R' → 'r', 'L' → 'l', 'L/R' → 'b' (both)
- */
-const STICKING_ENCODE: Record<string, string> = {
-  'R': 'r',
-  'L': 'l',
-  'L/R': 'b',
-};
-const STICKING_DECODE: Record<string, StickingValue> = {
-  'r': 'R',
-  'l': 'L',
-  'b': 'L/R',
-};
-/** Valid decoded sticking characters (excluding rest '-') */
-const VALID_STICKING_CHARS = new Set(['r', 'l', 'b', '-']);
-
-/** Valid divisions - matches Division type in types.ts */
-const VALID_DIVISIONS = [4, 8, 12, 16, 24, 32, 48] as const;
-const VALID_NOTE_VALUES = [4, 8, 16] as const;
-
-/** Zod schema for time signature string (e.g., "4/4") */
-const timeSignatureSchema = z.string()
-  .regex(/^\d{1,2}\/\d{1,2}$/, 'Invalid time signature format')
-  .transform((str): { beats: number; noteValue: number } => {
-    const [beats, noteValue] = str.split('/').map(Number);
-    return { beats, noteValue };
-  })
-  .refine(
-    (ts) => ts.beats >= VALIDATION.BEATS.MIN && ts.beats <= VALIDATION.BEATS.MAX,
-    { message: `Beats must be between ${VALIDATION.BEATS.MIN} and ${VALIDATION.BEATS.MAX}` }
-  )
-  .refine(
-    (ts) => VALID_NOTE_VALUES.includes(ts.noteValue as 4 | 8 | 16),
-    { message: `Note value must be one of: ${VALID_NOTE_VALUES.join(', ')}` }
-  );
-
-/** Zod schema for division */
-const divisionSchema = z.coerce.number()
-  .refine(
-    (n) => VALID_DIVISIONS.includes(n as Division),
-    { message: `Division must be one of: ${VALID_DIVISIONS.join(', ')}` }
-  )
-  .transform((n) => n as Division);
-
-/** Zod schema for tempo */
-const tempoSchema = z.coerce.number()
-  .min(VALIDATION.TEMPO.MIN, `Tempo must be at least ${VALIDATION.TEMPO.MIN}`)
-  .max(VALIDATION.TEMPO.MAX, `Tempo must be at most ${VALIDATION.TEMPO.MAX}`)
-  .default(VALIDATION.TEMPO.DEFAULT);
-
-/** Zod schema for swing */
-const swingSchema = z.coerce.number()
-  .min(VALIDATION.SWING.MIN, `Swing must be at least ${VALIDATION.SWING.MIN}`)
-  .max(VALIDATION.SWING.MAX, `Swing must be at most ${VALIDATION.SWING.MAX}`)
-  .default(VALIDATION.SWING.DEFAULT);
-
-/** Zod schema for measures count */
-const measuresSchema = z.coerce.number()
-  .min(VALIDATION.MEASURES.MIN, `Measures must be at least ${VALIDATION.MEASURES.MIN}`)
-  .max(VALIDATION.MEASURES.MAX, `Measures must be at most ${VALIDATION.MEASURES.MAX}`)
-  .default(VALIDATION.MEASURES.DEFAULT);
-
-/** Zod schema for text fields (title, author, comments) */
-const titleSchema = z.string()
-  .max(VALIDATION.TITLE_MAX_LENGTH, `Title must be at most ${VALIDATION.TITLE_MAX_LENGTH} characters`)
-  .optional();
-
-const authorSchema = z.string()
-  .max(VALIDATION.AUTHOR_MAX_LENGTH, `Author must be at most ${VALIDATION.AUTHOR_MAX_LENGTH} characters`)
-  .optional();
-
-const commentsSchema = z.string()
-  .max(VALIDATION.COMMENTS_MAX_LENGTH, `Comments must be at most ${VALIDATION.COMMENTS_MAX_LENGTH} characters`)
-  .optional();
-
-/** Zod schema for voice patterns - validates format and length */
-const patternSchema = z.string()
-  .max(VALIDATION.PATTERN_MAX_LENGTH, `Pattern too long`)
-  .regex(/^[|a-zA-Z0-9\-+^]*$/, 'Invalid characters in pattern')
-  .optional();
-
-/**
- * Zod schema for sticking parameter — validates character set and length (T-02-07)
- * Valid chars: 'r', 'l', 'b', '-', '|'
- */
-const stickingParamSchema = z.string()
-  .max(VALIDATION.STICKING_MAX_LENGTH, 'Sticking parameter too long')
-  .regex(/^[rlb|-]*$/, 'Invalid characters in sticking parameter')
-  .optional();
+import {
+  VALIDATION,
+  MEASURE_SEP,
+  REST_CHAR,
+  timeSignatureSchema,
+  divisionSchema,
+  tempoSchema,
+  swingSchema,
+  measuresSchema,
+  titleSchema,
+  authorSchema,
+  commentsSchema,
+  patternSchema,
+  stickingParamSchema,
+  safeParse,
+} from './urlCodecSchemas';
+import { encodeStickingPattern, decodeStickingPattern } from './stickingCodec';
 
 /** URL parameter names */
 const PARAM = {
@@ -150,12 +62,6 @@ const PARAM = {
   STICKING: 'Stk',
 } as const;
 
-/** Rest character in URL patterns */
-const REST_CHAR = '-';
-
-/** Measure separator in URL patterns */
-const MEASURE_SEP = '|';
-
 /**
  * Voice groups for URL encoding/decoding
  * Imported from centralized DrumVoiceConfig
@@ -168,7 +74,6 @@ const VOICE_GROUPS = URL_VOICE_GROUPS as Record<string, DrumVoice[]>;
  * Special handling for hihat group: encodes both hihat-closed and hihat-foot if present
  */
 function encodeVoicePattern(groove: GrooveData, voices: DrumVoice[]): string {
-  // Guard against empty voices array
   if (voices.length === 0) {
     return '';
   }
@@ -190,11 +95,9 @@ function encodeVoicePattern(groove: GrooveData, voices: DrumVoice[]): string {
       const hasHihatFoot = canEncodeHihatFoot && measureNotes['hihat-foot']?.[i];
 
       // Special case: hihat group with both closed and foot at same position.
-      // A lone hihat-foot still falls through to its own URL char below.
       if (hasHihatClosed && hasHihatFoot) {
-        char = '^'; // Combination character for both
+        char = '^';
       } else {
-        // Check each voice in priority order, take first match
         for (const voice of voices) {
           if (measureNotes[voice]?.[i]) {
             char = URL_TAB_CHARS[voice] || REST_CHAR;
@@ -220,16 +123,13 @@ function decodeVoicePattern(
   voices: DrumVoice[],
   notesPerMeasure: number
 ): Partial<Record<DrumVoice, boolean[]>> {
-  // Remove measure separators and get characters
   const chars = pattern.replace(/\|/g, '').split('');
   const result: Partial<Record<DrumVoice, boolean[]>> = {};
 
-  // Initialize all voices with false
   for (const voice of voices) {
     result[voice] = Array(notesPerMeasure).fill(false);
   }
 
-  // Build reverse lookup: char → voice
   const charToVoice: Record<string, DrumVoice> = {};
   for (const voice of voices) {
     const char = URL_TAB_CHARS[voice];
@@ -238,11 +138,9 @@ function decodeVoicePattern(
     }
   }
 
-  // Decode each position
   for (let i = 0; i < Math.min(chars.length, notesPerMeasure); i++) {
     const char = chars[i];
     if (char === '^') {
-      // Special combination character: both hihat-closed and hihat-foot
       if (result['hihat-closed']) result['hihat-closed']![i] = true;
       if (result['hihat-foot']) result['hihat-foot']![i] = true;
     } else if (char !== REST_CHAR) {
@@ -257,105 +155,26 @@ function decodeVoicePattern(
 }
 
 /**
- * Encode sticking arrays from all measures into a single URL-safe string.
- * Format: |<measure1chars>|<measure2chars>|...
- * Characters: 'r'=R, 'l'=L, 'b'=L/R, '-'=null
- * Returns null if all sticking values are null (nothing to encode).
- */
-function encodeStickingPattern(groove: GrooveData): string | null {
-  const hasAnySticking = groove.measures.some(m =>
-    m.sticking && m.sticking.some(v => v !== null)
-  );
-  if (!hasAnySticking) return null;
-
-  const parts: string[] = [];
-  for (const measure of groove.measures) {
-    const sticking = measure.sticking;
-    if (!sticking || sticking.length === 0) {
-      // Determine length from notes array to produce correct-length placeholders
-      const firstVoice = Object.keys(measure.notes)[0] as DrumVoice | undefined;
-      const length = firstVoice ? (measure.notes[firstVoice]?.length ?? 0) : 0;
-      parts.push('-'.repeat(length));
-    } else {
-      parts.push(sticking.map(v => (v !== null ? STICKING_ENCODE[v] : '-')).join(''));
-    }
-  }
-
-  return MEASURE_SEP + parts.join(MEASURE_SEP) + MEASURE_SEP;
-}
-
-/**
- * Decode sticking URL parameter back to per-measure StickingValue arrays.
- * Validates each character (T-02-07). Invalid characters are treated as null.
- * Returns an array of sticking arrays (one per measure), or null if no sticking param.
- */
-function decodeStickingPattern(
-  param: string,
-  numMeasures: number,
-  notesPerMeasure: number
-): (StickingValue[] | undefined)[] {
-  // Trim exactly one leading/trailing MEASURE_SEP so that empty middle measures
-  // (empty string segments) keep their positional index and don't get shifted (C8).
-  const trimmed = param.startsWith(MEASURE_SEP) ? param.slice(1) : param;
-  const trimmed2 = trimmed.endsWith(MEASURE_SEP) ? trimmed.slice(0, -1) : trimmed;
-  const measurePatterns = trimmed2.split(MEASURE_SEP); // no .filter — preserve empty slots
-  const result: (StickingValue[] | undefined)[] = [];
-
-  for (let m = 0; m < numMeasures; m++) {
-    const chars = measurePatterns[m] ?? '';
-    if (chars.length === 0) {
-      result.push(undefined);
-      continue;
-    }
-    const sticking: StickingValue[] = [];
-    for (let i = 0; i < notesPerMeasure; i++) {
-      const ch = i < chars.length ? chars[i] : '-';
-      if (!VALID_STICKING_CHARS.has(ch)) {
-        // T-02-07: reject invalid sticking characters — treat as null
-        sticking.push(null);
-      } else {
-        sticking.push(STICKING_DECODE[ch] ?? null);
-      }
-    }
-    // Only return a sticking array if it has at least one non-null value
-    const hasValues = sticking.some(v => v !== null);
-    result.push(hasValues ? sticking : undefined);
-  }
-
-  return result;
-}
-
-/**
  * Encode GrooveData to URL search params string
  */
 export function encodeGrooveToURL(groove: GrooveData): string {
   const params = new URLSearchParams();
 
-  // Basic params
   params.set(PARAM.TIME_SIG, `${groove.timeSignature.beats}/${groove.timeSignature.noteValue}`);
   params.set(PARAM.DIV, String(groove.division));
   params.set(PARAM.TEMPO, String(groove.tempo));
   params.set(PARAM.MEASURES, String(groove.measures.length));
-  
+
   if (groove.swing > 0) {
     params.set(PARAM.SWING, String(groove.swing));
   }
 
-  // Metadata (only include if non-empty)
-  if (groove.title) {
-    params.set(PARAM.TITLE, groove.title);
-  }
-  if (groove.author) {
-    params.set(PARAM.AUTHOR, groove.author);
-  }
-  if (groove.comments) {
-    params.set(PARAM.COMMENTS, groove.comments);
-  }
+  if (groove.title) params.set(PARAM.TITLE, groove.title);
+  if (groove.author) params.set(PARAM.AUTHOR, groove.author);
+  if (groove.comments) params.set(PARAM.COMMENTS, groove.comments);
 
-  // Use flattened notes to check if any voice has notes
   const flatNotes = getFlattenedNotes(groove);
 
-  // Voice patterns
   for (const [param, voices] of Object.entries(VOICE_GROUPS)) {
     const hasNotes = voices.some((v: DrumVoice) => flatNotes[v]?.some((n: boolean) => n));
     if (hasNotes) {
@@ -363,28 +182,12 @@ export function encodeGrooveToURL(groove: GrooveData): string {
     }
   }
 
-  // Sticking (per D-08): only include if any sticking values are set
   const stickingEncoded = encodeStickingPattern(groove);
   if (stickingEncoded !== null) {
     params.set(PARAM.STICKING, stickingEncoded);
   }
 
   return params.toString();
-}
-
-/**
- * Safe parse helper - returns default value on validation failure
- */
-function safeParse<T>(
-  schema: z.ZodType<T>,
-  value: string | null | undefined,
-  defaultValue: T
-): T {
-  if (value === null || value === undefined) {
-    return defaultValue;
-  }
-  const result = schema.safeParse(value);
-  return result.success ? result.data : defaultValue;
 }
 
 /**
@@ -397,7 +200,6 @@ export function decodeURLToGroove(urlOrParams: string | URLSearchParams): Groove
     ? new URLSearchParams(urlOrParams.startsWith('?') ? urlOrParams.slice(1) : urlOrParams)
     : urlOrParams;
 
-  // Parse and validate basic params with Zod schemas
   const timeSigResult = timeSignatureSchema.safeParse(params.get(PARAM.TIME_SIG) || '4/4');
   const timeSignature: TimeSignature = timeSigResult.success
     ? { beats: timeSigResult.data.beats, noteValue: timeSigResult.data.noteValue as 4 | 8 | 16 }
@@ -408,7 +210,6 @@ export function decodeURLToGroove(urlOrParams: string | URLSearchParams): Groove
   const swing = safeParse(swingSchema, params.get(PARAM.SWING), VALIDATION.SWING.DEFAULT);
   const numMeasures = safeParse(measuresSchema, params.get(PARAM.MEASURES), VALIDATION.MEASURES.DEFAULT);
 
-  // Parse and validate metadata with length limits
   const title = safeParse(titleSchema, params.get(PARAM.TITLE), undefined);
   const author = safeParse(authorSchema, params.get(PARAM.AUTHOR), undefined);
   const comments = safeParse(commentsSchema, params.get(PARAM.COMMENTS), undefined);
@@ -417,23 +218,20 @@ export function decodeURLToGroove(urlOrParams: string | URLSearchParams): Groove
   // e.g., 4/4 with 16ths: (16/4) * 4 = 16
   const notesPerMeasure = (division / timeSignature.noteValue) * timeSignature.beats;
 
-  // Start with empty notes for each measure
   const measures: { notes: Record<DrumVoice, boolean[]>; sticking?: StickingValue[] }[] = [];
   for (let m = 0; m < numMeasures; m++) {
     measures.push({ notes: createEmptyNotesRecord(notesPerMeasure) });
   }
 
-  // Decode voice patterns with validation
   for (const [param, voices] of Object.entries(VOICE_GROUPS)) {
     const rawPattern = params.get(param);
-    // Validate pattern format before processing
     const pattern = safeParse(patternSchema, rawPattern, undefined);
     if (pattern) {
       // Trim exactly one leading/trailing MEASURE_SEP then split without filtering empties,
       // so that a future empty-middle-measure segment keeps its positional index (C8).
       const trimmed = pattern.startsWith(MEASURE_SEP) ? pattern.slice(1) : pattern;
       const trimmed2 = trimmed.endsWith(MEASURE_SEP) ? trimmed.slice(0, -1) : trimmed;
-      const measurePatterns = trimmed2.split(MEASURE_SEP); // no .filter — preserve empty slots
+      const measurePatterns = trimmed2.split(MEASURE_SEP);
 
       for (let m = 0; m < Math.min(measurePatterns.length, numMeasures); m++) {
         const decoded = decodeVoicePattern(
@@ -482,11 +280,8 @@ export function decodeURLToGroove(urlOrParams: string | URLSearchParams): Groove
  * - Error threshold: 8000 chars (likely to fail)
  */
 export const URL_LENGTH_LIMITS = {
-  /** Safe URL length for all browsers/servers */
   SAFE: 2000,
-  /** Warning threshold - consider compression */
   WARNING: 1500,
-  /** Maximum before likely failure */
   MAX: 8000,
 } as const;
 
@@ -540,10 +335,6 @@ export function validateURLLength(url: string): URLValidationResult {
  * in production deployments (e.g., /scribe/ or /groovy/).
  */
 export function getShareableURL(groove: GrooveData, baseURL?: string, mode: 'embed' | 'editor' = 'embed'): string {
-  // Use provided baseURL, or construct from origin + Vite's BASE_URL
-  // import.meta.env.BASE_URL is set by Vite based on the 'base' config:
-  // - Development: '/'
-  // - Production: '/scribe/' (or whatever PRODUCTION_BASE_PATH is set to)
   const base = baseURL || window.location.origin + (import.meta.env.BASE_URL || '/');
   const params = encodeGrooveToURL(groove);
   const url = `${base}?${params}`;
@@ -552,7 +343,6 @@ export function getShareableURL(groove: GrooveData, baseURL?: string, mode: 'emb
 
 /**
  * Get shareable URL with validation
- * Returns the URL along with validation status
  */
 export function getShareableURLWithValidation(
   groove: GrooveData,
@@ -594,6 +384,5 @@ export const GrooveURLCodec = {
   validateURLLength,
   hasGrooveParams,
   URL_LENGTH_LIMITS,
-  /** Validation constraints for URL parameters */
   VALIDATION,
 };
